+++ react/containers/Dropzone/view.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import { selectFile, updateError } from '../actions';
-import { validateFile } from '../utils/file.js';
-import Preview from '../components/Preview.jsx';
+import { validateFile } from '../../utils/file.js';
+import Preview from '../../components/Preview.jsx';
 
 class Dropzone extends React.Component {
   constructor (props) {
@@ -139,24 +137,4 @@ class Dropzone extends React.Component {
   }
 };
 
-const mapStateToProps = state => {
-  return {
-    file     : state.file,
-    thumbnail: state.metadata.thumbnail,
-    fileError: state.error.file,
-  };
-};
-
-const mapDispatchToProps = dispatch => {
-  return {
-    onFileSelect: (file) => {
-      dispatch(selectFile(file));
-      dispatch(updateError('publishSubmit', null));
-    },
-    onFileError: (value) => {
-      dispatch(updateError('file', value));
-    },
-  };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(Dropzone);
+export default Dropzone;
