+++ react/containers/PublishThumbnailInput/index.js
@@ -1,0 +1,19 @@
+import {connect} from 'react-redux';
+import {updateMetadata} from '../../actions';
+import View from './view.jsx';
+
+const mapStateToProps = state => {
+  return {
+    thumbnail: state.metadata.thumbnail,
+  };
+};
+
+const mapDispatchToProps = dispatch => {
+  return {
+    onThumbnailChange: (name, value) => {
+      dispatch(updateMetadata(name, value));
+    },
+  };
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(View);
